requirements.txt
plyfile
opencv-python
trimesh==2.35.39
tensorboardX
easydict
tqdm
h5py
matplotlib
elevate
numpy-quaternion
imageio
pandas
seaborn